use size_t where can, update where neede

+++ src/parser_data.h
@@ -22,13 +22,13 @@ struct Init_Segment {
   union {
     int32_t value; // number
     struct {
-      int count;                              // how many
+      size_t count;                           // how many
       int32_t value;                          // number - or ? if is_uninit
     } dup;                                    // dup
     char string[MAX_INIT_SEGMENT_STRING_LEN]; // string
   } data;
-  int element_count; // length of string/count in dup/1 for number
-  int is_uninit;     // is value/dup un-initialized
+  size_t element_count; // length of string/count in dup/1 for number
+  int is_uninit;        // is value/dup un-initialized
 };
 
 // when delaring an assembler variable
@@ -39,7 +39,7 @@ struct Data_Declaration {
   struct Init_Segment *segments; // array of segments
   size_t segment_count;
 
-  int total_size;      // total size of all segments->element_count
+  size_t total_size;   // total size of all segments->element_count
   int is_fully_uninit; // if 1 if and only if every init segment is_uninit
 };
